requirements.txt
django
pillow
django-autoslug
django-widget-tweaks
django-localflavor
openpyxl
whitenoise
django-model-utils
django-crispy-forms
gunicorn
mysqlclient
requests
